+++ codigo/Testes/UsoDeVagaTest.java
@@ -4,6 +4,8 @@ import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
 
+import main.Vaga;
+
 public class UsoDeVagaTest {
 
     @Test
